+++ fabric/src/main/java/dev/muon/questkilltask/mixin/FTBQuestsEventHandlerMixin.java
@@ -1,13 +1,18 @@
 package dev.muon.questkilltask.mixin;
 
 import dev.architectury.event.EventResult;
+import dev.architectury.hooks.level.entity.PlayerHooks;
 import dev.ftb.mods.ftbquests.FTBQuestsEventHandler;
 import dev.ftb.mods.ftbquests.quest.ServerQuestFile;
 import dev.ftb.mods.ftbquests.quest.TeamData;
 import dev.ftb.mods.ftbquests.quest.task.KillTask;
 import dev.muon.questkilltask.DamageTracker;
+import dev.muon.questkilltask.QuestKillTask;
+import dev.muon.questkilltask.QuestKillTaskFabric;
 import dev.muon.questkilltask.QuestProcessor;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -20,9 +25,6 @@ import java.util.*;
 
 @Mixin(value = FTBQuestsEventHandler.class, remap = false)
 public class FTBQuestsEventHandlerMixin {
-    @Shadow(remap = false)
-    private List<KillTask> killTasks;
-
     @Unique
     private QuestProcessor questKillTask$questProcessor;
 
@@ -38,10 +40,14 @@ public class FTBQuestsEventHandlerMixin {
      * It's unlikely other mods interact with this task.
      * This can be rewritten if issues arise.
      */
+
+    // TODO: Clean up, some duplicate logic is being run
     @Inject(method = "playerKill", at = @At("HEAD"), cancellable = true)
     private void onPlayerKill(LivingEntity entity, DamageSource source, CallbackInfoReturnable<EventResult> cir) {
+        if ((entity.level().isClientSide) || (source.getEntity() != null && source.getEntity().level().isClientSide)) return;
+
         if (questKillTask$questProcessor == null) {
-            questKillTask$questProcessor = new QuestProcessor(killTasks);
+            questKillTask$questProcessor = new QuestProcessor();
         }
 
         if (!questKillTask$questProcessor.shouldProcessKill(entity)) {
@@ -51,12 +57,12 @@ public class FTBQuestsEventHandlerMixin {
 
         ServerQuestFile questFile = questKillTask$questProcessor.getQuestFile();
         if (questFile == null) {
+            QuestKillTask.LOG.warn("Unable to retrieve Server Quest File!");
             cir.setReturnValue(EventResult.pass());
             return;
         }
 
-        TeamData killerTeam = questKillTask$questProcessor.getKillerTeam(source, questFile);
-        questKillTask$questProcessor.processDamagingTeams(entity, questFile, killerTeam);
+        questKillTask$questProcessor.processDamagingTeams(entity, questFile);
 
         DamageTracker.clearEntityTracking(entity);
         cir.setReturnValue(EventResult.pass());