+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# 0.1.0
+## 0.2.0
+- Fix an issue with entity death checks that could cause kill tasks to attempt to be processed on the client, and lead to frozen quest progress
+- Fix an issue with task caching that could cause the kill task list to desync between the server and the client
+
+## 0.1.0
 - Initial Release